last push for web server and optimizations to recognition
last push that wasnt made during the hackathon

+++ tts.py
@@ -5,9 +5,13 @@ from io import BytesIO
 import uuid, os
 
 def speak(text: str, lang="en"):
-    tts = gTTS(text, lang=lang)
+    if text == "" or text is None:
+        return
+
+    tts = gTTS(text, lang=lang, tld="com")
     filename = f"/tmp/{uuid.uuid4()}.mp3"
     tts.save(filename)
     
     playsound(filename)
+    print("audio end")
     os.remove(filename)
